refactor(run): split shuck-run into modules
